+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ensure temporary file edited during `arcanus edit` ends with `.yaml` so
   editor chooses correct syntax highlighter
+* Fix password setting to not include trailing newline
 
 ## 0.3.0