+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Arcanus Changelog
 
-## master (unreleased)
+## 0.4.0
 
 * Ensure temporary file edited during `arcanus edit` ends with `.yaml` so
   editor chooses correct syntax highlighter